requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn==0.27.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Pub/Sub
google-cloud-pubsub>=2.21.0

# Groq
groq>=0.9.0

# Environment variables
python-dotenv==1.0.0

# Excel
openpyxl==3.1.2

pyyaml==6.0.3

